# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurant",
        "description": "REST API para autenticación de usuarios, administración del menú y gestión de ingredientes.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/pizzas/{pizza}/ingredients": {
            "get": {
                "tags": [
                    "Pizzas"
                ],
                "summary": "Lista los ingredientes asociados a una pizza",
                "operationId": "db9dfc3b7c11eeb0173b556f5c5eadf7",
                "parameters": [
                    {
                        "name": "pizza",
                        "in": "path",
                        "description": "ID de la pizza",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de ingredientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PizzaIngredientCollectionResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pizzas"
                ],
                "summary": "Asocia un ingrediente a la pizza",
                "operationId": "1eb28cb7a9d37adba7b2be953559f0d3",
                "parameters": [
                    {
                        "name": "pizza",
                        "in": "path",
                        "description": "ID de la pizza",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ingredient_id"
                                ],
                                "properties": {
                                    "ingredient_id": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ingrediente vinculado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PizzaIngredientResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pizzas/{pizza}/ingredients/{pivot}": {
            "delete": {
                "tags": [
                    "Pizzas"
                ],
                "summary": "Elimina la relación entre una pizza y un ingrediente",
                "operationId": "2b8433c749ddabf1aa5df7e8d52412cd",
                "parameters": [
                    {
                        "name": "pizza",
                        "in": "path",
                        "description": "ID de la pizza",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pivot",
                        "in": "path",
                        "description": "ID del registro pizza_ingredient",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ingrediente desvinculado"
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IngredientImage": {
                "properties": {
                    "public_id": {
                        "type": "string",
                        "example": "api-restaurant/ingredients/abc123"
                    },
                    "secure_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://res.cloudinary.com/demo/image/upload/v1/ingredients.png"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "format": {
                        "type": "string",
                        "example": "png"
                    },
                    "resource_type": {
                        "type": "string",
                        "example": "image"
                    },
                    "bytes": {
                        "type": "integer",
                        "example": 450123
                    },
                    "width": {
                        "type": "integer",
                        "example": 1200
                    },
                    "height": {
                        "type": "integer",
                        "example": 800
                    }
                },
                "type": "object",
                "nullable": true
            },
            "IngredientResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "type": "string",
                        "example": "Mozzarella"
                    },
                    "price": {
                        "type": "string",
                        "example": "2.50"
                    },
                    "ingredient": {
                        "type": "string",
                        "example": "cheese"
                    },
                    "available": {
                        "type": "boolean",
                        "example": true
                    },
                    "image_url": {
                        "$ref": "#/components/schemas/IngredientImage"
                    }
                },
                "type": "object"
            },
...
            "PizzaIngredientResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 501
                    },
                    "pizza_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "ingredient_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "ingredient": {
                        "$ref": "#/components/schemas/IngredientResource"
                    }
                },
                "type": "object"
            },
            "PizzaIngredientResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "pizza_ingredient": {
                                "$ref": "#/components/schemas/PizzaIngredientResource"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PizzaIngredientCollectionResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "ingredients": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/PizzaIngredientResource"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "Las credenciales no coinciden."
                            ]
                        },
                        "additionalProperties": {
                            "schema": "Schemas",
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "UnauthorizedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    },
                    "status": {
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
            "ForbiddenError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    },
                    "status": {
                        "type": "integer",
                        "example": 403
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Incluye el token generado en el login o registro usando el formato: Bearer {token}",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}